+++ VoiceCraft.Core/Network/Packets/SetNamePacket.cs
@@ -5,11 +5,11 @@ namespace VoiceCraft.Core.Network.Packets
     public class SetNamePacket : VoiceCraftPacket
     {
         public override PacketType PacketType => PacketType.SetName;
-        public byte Id { get; private set; }
+        public int Id { get; private set; }
         public string Name { get; private set; }
         
 
-        public SetNamePacket(byte id = 0, string name = "")
+        public SetNamePacket(int id = 0, string name = "")
         {
             Id = id;
             Name = name;
@@ -23,7 +23,7 @@ namespace VoiceCraft.Core.Network.Packets
 
         public override void Deserialize(NetDataReader reader)
         {
-            Id = reader.GetByte();
+            Id = reader.GetInt();
             Name = reader.GetString(Constants.MaxStringLength);
         }
     }
